doubled (hehe) awall precision

+++ Castle/AutoWall.cpp
@@ -1,26 +1,26 @@
 #include "autowall.h"
 #include "Math.h"
-float Autowall::GetHitgroupDamageMultiplier(HitGroups iHitGroup)
+double Autowall::GetHitgroupDamageMultiplier(HitGroups iHitGroup)
 {
 	switch (iHitGroup)
 	{
 	case HitGroups::HITGROUP_HEAD:
-		return 4.0f;
+		return 4.0;
 	case HitGroups::HITGROUP_CHEST:
 	case HitGroups::HITGROUP_LEFTARM:
 	case HitGroups::HITGROUP_RIGHTARM:
-		return 1.0f;
+		return 1.0;
 	case HitGroups::HITGROUP_STOMACH:
-		return 1.25f;
+		return 1.25;
 	case HitGroups::HITGROUP_LEFTLEG:
 	case HitGroups::HITGROUP_RIGHTLEG:
-		return 0.75f;
+		return 0.75;
 	default:
-		return 1.0f;
+		return 1.0;
 	}
 }
 
-void Autowall::ScaleDamage(HitGroups hitgroup, C_BasePlayer* enemy, float weapon_armor_ratio, float& current_damage)
+void Autowall::ScaleDamage(HitGroups hitgroup, C_BasePlayer* enemy, double weapon_armor_ratio, double& current_damage)
 {
 	current_damage *= Autowall::GetHitgroupDamageMultiplier(hitgroup);
 
@@ -29,20 +29,20 @@ void Autowall::ScaleDamage(HitGroups hitgroup, C_BasePlayer* enemy, float weapon
 		if (hitgroup == HitGroups::HITGROUP_HEAD)
 		{
 			if (enemy->HasHelmet())
-				current_damage *= weapon_armor_ratio * 0.75f;
+				current_damage *= weapon_armor_ratio * 0.75;
 		}
 		else
-			current_damage *= weapon_armor_ratio * 0.5f;
+			current_damage *= weapon_armor_ratio * 0.5;
 	}
 }
 
 bool Autowall::TraceToExit(Vector& end, trace_t* enter_trace, Vector start, Vector dir, trace_t* exit_trace)
 {
-	float distance = 0.0f;
+	double distance = 0.0;
 
-	while (distance <= 90.0f)
+	while (distance <= 90.0)
 	{
-		distance += 4.0f;
+		distance += 4.0;
 		end = start + dir * distance;
 
 		auto point_contents = pTrace->GetPointContents(end, MASK_SHOT_HULL | CONTENTS_HITBOX, NULL);
@@ -50,7 +50,7 @@ bool Autowall::TraceToExit(Vector& end, trace_t* enter_trace, Vector start, Vect
 		if (point_contents & MASK_SHOT_HULL && !(point_contents & CONTENTS_HITBOX))
 			continue;
 
-		auto new_end = end - (dir * 4.0f);
+		auto new_end = end - (dir * 4.0);
 
 		Ray_t ray;
 		ray.Init(end, new_end);
@@ -65,7 +65,7 @@ bool Autowall::TraceToExit(Vector& end, trace_t* enter_trace, Vector start, Vect
 
 			pTrace->TraceRay(ray, 0x600400B, &filter, exit_trace);
 
-			if ((exit_trace->fraction < 1.0f || exit_trace->allsolid) && !exit_trace->startsolid)
+			if ((exit_trace->fraction < 1.0 || exit_trace->allsolid) && !exit_trace->startsolid)
 			{
 				end = exit_trace->endpos;
 				return true;
@@ -88,9 +88,9 @@ bool Autowall::TraceToExit(Vector& end, trace_t* enter_trace, Vector start, Vect
 		if (exit_trace->surface.flags >> 7 & 1 && !(enter_trace->surface.flags >> 7 & 1))
 			continue;
 
-		if (exit_trace->plane.normal.Dot(dir) <= 1.0f)
+		if (exit_trace->plane.normal.Dot(dir) <= 1.0)
 		{
-			auto fraction = exit_trace->fraction * 4.0f;
+			auto fraction = exit_trace->fraction * 4.0;
 			end = end - (dir * fraction);
 
 			return true;
@@ -104,12 +104,12 @@ bool Autowall::HandleBulletPenetration(CCSWeaponInfo* weaponInfo, FireBulletData
 {
 	surfacedata_t *enter_surface_data = pPhysics->GetSurfaceData(data.enter_trace.surface.surfaceProps);
 	int enter_material = enter_surface_data->game.material;
-	float enter_surf_penetration_mod = *(float*)((uint8_t*)enter_surface_data + 76);
+	double enter_surf_penetration_mod = *(float*)((uint8_t*)enter_surface_data + 76);
 
 	data.trace_length += data.enter_trace.fraction * data.trace_length_remaining;
-	data.current_damage *= powf(weaponInfo->GetRangeModifier(), data.trace_length * 0.002f);
+	data.current_damage *= powf(weaponInfo->GetRangeModifier(), data.trace_length * 0.002);
 
-	if (data.trace_length > 3000.f || enter_surf_penetration_mod < 0.1f)
+	if (data.trace_length > 3000.0 || enter_surf_penetration_mod < 0.1)
 		data.penetrate_count = 0;
 
 	if (data.penetrate_count <= 0)
@@ -124,43 +124,43 @@ bool Autowall::HandleBulletPenetration(CCSWeaponInfo* weaponInfo, FireBulletData
 	surfacedata_t *exit_surface_data = pPhysics->GetSurfaceData(trace_exit.surface.surfaceProps);
 	int exit_material = exit_surface_data->game.material;
 
-	float exit_surf_penetration_mod = *(float*)((uint8_t*)exit_surface_data + 76);
-	float final_damage_modifier = 0.16f;
-	float combined_penetration_modifier = 0.0f;
+	double exit_surf_penetration_mod = *(float*)((uint8_t*)exit_surface_data + 76);
+	double final_damage_modifier = 0.16;
+	double combined_penetration_modifier = 0.0;
 
 	if ((data.enter_trace.contents & CONTENTS_GRATE) != 0 || enter_material == 89 || enter_material == 71)
 	{
-		combined_penetration_modifier = 3.0f;
-		final_damage_modifier = 0.05f;
+		combined_penetration_modifier = 3.0;
+		final_damage_modifier = 0.05;
 	}
 	else
-		combined_penetration_modifier = (enter_surf_penetration_mod + exit_surf_penetration_mod) * 0.5f;
+		combined_penetration_modifier = (enter_surf_penetration_mod + exit_surf_penetration_mod) * 0.5;
 
 	if (enter_material == exit_material)
 	{
 		if (exit_material == 87 || exit_material == 85)
-			combined_penetration_modifier = 3.0f;
+			combined_penetration_modifier = 3.0;
 		else if (exit_material == 76)
-			combined_penetration_modifier = 2.0f;
+			combined_penetration_modifier = 2.0;
 	}
 
-	float v34 = fmaxf(0.f, 1.0f / combined_penetration_modifier);
-	float v35 = (data.current_damage * final_damage_modifier) + v34 * 3.0f * fmaxf(0.0f, (3.0f / weaponInfo->GetPenetration()) * 1.25f);
-	float thickness = (trace_exit.endpos - data.enter_trace.endpos).Length();
+	double v34 = fmaxf(0.0, 1.0 / combined_penetration_modifier);
+	double v35 = (data.current_damage * final_damage_modifier) + v34 * 3.0 * fmaxf(0.0, (3.0 / weaponInfo->GetPenetration()) * 1.25);
+	double thickness = (trace_exit.endpos - data.enter_trace.endpos).Length();
 
 	thickness *= thickness;
 	thickness *= v34;
-	thickness /= 24.0f;
+	thickness /= 24.0;
 
-	float lost_damage = fmaxf(0.0f, v35 + thickness);
+	double lost_damage = fmaxf(0.0f, v35 + thickness);
 
 	if (lost_damage > data.current_damage)
 		return false;
 
-	if (lost_damage >= 0.0f)
+	if (lost_damage >= 0.0)
 		data.current_damage -= lost_damage;
 
-	if (data.current_damage < 1.0f)
+	if (data.current_damage < 1.0)
 		return false;
 
 	data.src = trace_exit.endpos;
@@ -185,10 +185,10 @@ bool Autowall::SimulateFireBullet(C_BaseCombatWeapon* pWeapon, bool teamCheck, F
 	CCSWeaponInfo* weaponInfo = pWeapon->GetCSWpnData();
 
 	data.penetrate_count = 4;
-	data.trace_length = 0.0f;
+	data.trace_length = 0.0;
 	data.current_damage = (float)weaponInfo->GetDamage();
 
-	while (data.penetrate_count > 0 && data.current_damage >= 1.0f)
+	while (data.penetrate_count > 0 && data.current_damage >= 1.0)
 	{
 		data.trace_length_remaining = weaponInfo->GetRange() - data.trace_length;
 
@@ -199,21 +199,21 @@ bool Autowall::SimulateFireBullet(C_BaseCombatWeapon* pWeapon, bool teamCheck, F
 		TraceLine(data.src, end, MASK_SHOT, localplayer, &data.enter_trace);
 
 		Ray_t ray;
-		ray.Init(data.src, end + data.direction * 40.f);
+		ray.Init(data.src, end + data.direction * 40.0);
 
 		pTrace->TraceRay(ray, MASK_SHOT, &data.filter, &data.enter_trace);
 		
 		// real man traces
 		TraceLine(data.src, end, MASK_SHOT, localplayer, &data.enter_trace);
-		TraceLine(data.src, end + data.direction * 40.f, MASK_SHOT, localplayer, &data.enter_trace);
+		TraceLine(data.src, end + data.direction * 40.0, MASK_SHOT, localplayer, &data.enter_trace);
 
-		if (data.enter_trace.fraction == 1.0f)
+		if (data.enter_trace.fraction == 1.0)
 			break;
 
 		if (data.enter_trace.hitgroup <= HitGroups::HITGROUP_RIGHTLEG && data.enter_trace.hitgroup > HitGroups::HITGROUP_GENERIC)
 		{
 			data.trace_length += data.enter_trace.fraction * data.trace_length_remaining;
-			data.current_damage *= powf(weaponInfo->GetRangeModifier(), data.trace_length * 0.002f);
+			data.current_damage *= powf(weaponInfo->GetRangeModifier(), data.trace_length * 0.002);
 
 			C_BasePlayer* player = (C_BasePlayer*)data.enter_trace.m_pEnt;
 			if (teamCheck && player->GetTeam() == localplayer->GetTeam())
@@ -231,9 +231,9 @@ bool Autowall::SimulateFireBullet(C_BaseCombatWeapon* pWeapon, bool teamCheck, F
 	return false;
 }
 
-float Autowall::GetDamage(const Vector& point, bool teamCheck, FireBulletData& fData)
+double Autowall::GetDamage(const Vector& point, bool teamCheck, FireBulletData& fData)
 {
-	float damage = 0.f;
+	double damage = 0.0;
 	Vector dst = point;
 	C_BasePlayer* localplayer = (C_BasePlayer*)pEntityList->GetClientEntity(pEngine->GetLocalPlayer());
 	FireBulletData data;
